Added tests, not working yet
